I created a new style.css file and a new page for my website which is Champions.html. In the css file i transferred all my css into the new style.css file. In the champions page I added a
temporary dropdown menu to show all the lanes of league of legends, the popular builds and the gamemodes for league of legends.
In terms of feedback I had receieved from my classmates, they said that I should change the colours of the champions.html page to make it more visually appealing towards the target audience, maybe to change the colour from red to something like a gradient of blue.